Adds Device.RepeatState and Device.SetRepeatState methods. These allow us to get/set the repeat state of a device.
The values indicate (in milliseconds) the delay before
the device starts repeating and the delay between
subsequent repeats. This might apply to a keyboard where
the user presses and holds a key.

E.g.: We see an initial character immediately, then
another @initial milliseconds later and after that,
once every @subsequent milliseconds, until the key
is released.