Adds the evdev.Find() function, along with IsKeyboard, IsMouse and IsJoystick. Find() can be used to return a list of attached devices which qualify as the device represented by the given constant. E.g.:
    keyboards, err := evdev.Find(evdev.Keyboard)

This returns a list of open devices which behave like a keyboard.
These may not actually be keyboards, but they support the events
that we expect from one: reporting key states and LED events.

When looking for Mouse devices, this may also return trackpads
and multi-touch screens. it is up to the host application to
figure out which one to actually use.

Adds an example program to demonstrate its use.